requirements.txt
appdirs==1.4.4
argcomplete==3.5.2
asgiref==3.8.1
attrs==24.3.0
autobahn==24.4.2
Automat==24.8.1
certifi==2024.8.30
cffi==1.17.1
channels==4.2.0
chardet==5.2.0
charset-normalizer==3.3.2
colorama==0.4.6
constantly==23.10.4
cryptography==43.0.3
daphne==4.1.2
Django==5.0.9
django-cors-headers==4.4.0
django-eventstream==5.3.1
django-grip==3.5.1
djangorestframework==3.15.2
gripcontrol==4.2.0
hyperlink==21.0.0
idna==3.10
importlib_resources==6.4.5
incremental==24.7.2
MarkupSafe==3.0.2
msgpack==1.1.0
mssql-django==1.5
numpy==2.2.1
pandas==2.2.3
pillow==11.0.0
pubcontrol==3.5.0
pyasn1==0.6.1
pyasn1_modules==0.4.1
pycparser==2.22
pycryptodomex==3.21.0
PyJWT==2.10.1
pymssql==2.3.1
pyodbc==5.1.0
pyOpenSSL==24.3.0
pyserial==3.5
python-barcode==0.15.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-escpos==3.1
pytz==2024.2
PyYAML==6.0.2
qrcode==8.0
reportlab==4.2.5
requests==2.32.3
scrapetube==2.5.1
service-identity==24.2.0
setuptools==75.6.0
six==1.17.0
sqlparse==0.5.1
Twisted==24.11.0
txaio==23.1.1
typing_extensions==4.12.2
tzdata==2024.1
urllib3==2.2.3
waitress==3.0.2
Werkzeug==3.1.3
zope.interface==7.2
